refactor tests into single file
